+++ .eslintrc.json
@@ -6,7 +6,8 @@
     },
     "extends": [
         "plugin:prettier/recommended",
-        "plugin:react/recommended"
+        "plugin:react/recommended",
+        "plugin:@typescript-eslint/recommended"
     ],
     "globals": {
         "Atomics": "readonly",
@@ -23,7 +24,7 @@
     },
     "plugins": ["react"],
     "rules": {
-        "prettier/prettier": "warn",
+        "prettier/prettier": "off",
         "react/react-in-jsx-scope": "off",
         "react/prop-types": "warn"
     }
