+++ pages/_app.js
@@ -1,3 +1,5 @@
+/* eslint @typescript-eslint/explicit-function-return-type: 0 */
+
 import React from "react";
 import App from "next/app";
 import Head from "next/head";
