--- pages/home.js
@@ -1,13 +1,0 @@
-import React, { Component } from "react";
-
-class Home extends Component {
-  render() {
-    return (
-      <div>
-        <h1>Hello hhh!!!</h1>
-      </div>
-    );
-  }
-}
-
-export default Home;
