+++ pages/home.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+export default class Home extends React.Component {
+  render() {
+    return (
+      <div>
+        <h1>Hello hhh!!!</h1>
+      </div>
+    );
+  }
+}
