--- pages/label.js
@@ -1,13 +1,0 @@
-import React, { Component } from "react";
-
-class Label extends Component {
-  render() {
-    return (
-      <div>
-        <h3>This is a label</h3>
-      </div>
-    );
-  }
-}
-
-export default Label;
