+++ pages/label.tsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+export default class Label extends React.Component {
+  render() {
+    return (
+      <h3>This is a label</h3>
+    );
+  }
+}
