+++ pages/post.tsx
@@ -1,4 +1,4 @@
-import React, { Component } from "react";
+import React from "react";
 import { useRouter } from "next/router";
 
 const Post = () => {
